requirements.txt
anthropic>=0.52.0
pydantic>=2.0
python-dotenv
